Move common objects introduced for #58 into lib
Moving things that are likely to be used across multiple OpenLIFU
modules.
